Added error message to restore command if message was an uploaded file

+++ cogs/mod.py
@@ -155,8 +155,8 @@ class Moderation:
     @commands.guild_only()
     async def restore(self, ctx):
         """**Must have the "Manage Messages" permission**
-        Restores the last deleted message"""
-
+        Restores the last deleted message. I can't restore uploaded files though, unfortunately
+        """
         if not ctx.author.permissions_in(ctx.channel).manage_messages:
             await ctx.send("You need the Manage Messages permission in order to use this command",
                            delete_after=5.0)
@@ -168,6 +168,13 @@ class Moderation:
         except:
             await ctx.send(
                 "Unable to find the last deleted message. Sorry!", delete_after=5.0)
+            await asyncio.sleep(5)
+            return await ctx.message.delete()
+
+        if last_delete['content'] == "":
+            await ctx.send(
+                "I can't restore this message. It was probably a file or something, and I can't "
+                "restore uploaded files. Sorry!", delete_after=5.0)
             await asyncio.sleep(5)
             return await ctx.message.delete()
